Moved filename also to NCInputTextBase
