+++ SOURCECONF.cmake
@@ -44,7 +44,7 @@ SET( ${TARGETLIB}_SOURCES
   NCPushButton.cc
   NCDumbTab.cc
   NCRadioButton.cc
-  NCInputText.cc
+  NCInputTextBase.cc
   NCInputField.cc
   NCDateField.cc
   NCTimeField.cc
@@ -118,7 +118,7 @@ SET( ${TARGETLIB}_HEADERS
   NCPushButton.h
   NCDumbTab.h
   NCRadioButton.h
-  NCInputText.h
+  NCInputTextBase.h
   NCInputField.h
   NCTimeField.h
   NCDateField.h
