+++ src/NCInputTextBase.cc
@@ -28,7 +28,7 @@
 #define  YUILogComponent "ncurses"
 #include <yui/YUILog.h>
 #include "NCurses.h"
-#include "NCInputText.h"
+#include "NCInputTextBase.h"
 
 #include <wctype.h>		// iswalnum()
