+++ src/NCInputTextBase.h
